+++ app/psu/page.tsx
@@ -1,7 +1,9 @@
+"use client";
+
 import PartsTab from "@/components/PartsTab";
 import { ThemeDropdown } from "@/components/ThemeDropdown";
 import TotalPrice from "@/components/TotalPrice";
-import { Case as CaseEntity } from "@/db/Case";
+import { Psu as PsuEntity } from "@/db/Psu";
 import pages from "@/utils/pages.json";
 import { Anchor, AppShell, Burger, Group, NavLink } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
@@ -29,7 +31,7 @@ function useBuf() {
 
 const pageList = pages;
 
-export default function Case() {
+export default function Psu() {
   const { buf } = useBuf();
   const [sql, setSql] = useState<initSqlJs.SqlJsStatic>();
   const [dataSource, setDatasource] = useState<DataSource>();
@@ -51,7 +53,7 @@ export default function Case() {
 
         const dataSource = new DataSource({
           type: "sqljs",
-          entities: [CaseEntity],
+          entities: [PsuEntity],
           database: new Uint8Array(buf),
         });
 
@@ -97,7 +99,7 @@ export default function Case() {
       </AppShell.Header>
       <AppShell.Navbar p="md">
         {pageList.map((x) => {
-          if (x.key === "case") {
+          if (x.key === "psu") {
             return <NavLink key={x.key} label={x.name} active />;
           } else {
             return (
@@ -118,7 +120,7 @@ export default function Case() {
         />
       </AppShell.Navbar>
       <AppShell.Main>
-        <PartsTab type="case" dataSource={dataSource} />
+        <PartsTab type="psu" dataSource={dataSource} />
       </AppShell.Main>
     </AppShell>
   );
